Changes made in Scikitlearn python scripts and java class files.(Numpy float64 which wasn't supported by the python executioner so changed to .astype(float))
Deleted the unused Json files

+++ konduit-serving-test/src/test/resources/scripts/scikitlearn/JsonScikitNDArrayInf.py
@@ -44,6 +44,6 @@ with open(filename, 'rb') as file:
  # Calculate the accuracy score and predict target values
 #score = pickle_model.score(Xtest, Ytest)
 #print("Test score: {0:.2f} %".format(100 * score))
-Ypredict = pickle_model.predict(Xr)
+Ypredict = pickle_model.predict(Xr).astype('float')
 
 print(Ypredict)
